feat: split recipient logic from escalation_tool file + update all tests
