+++ tests/agent/guardrails/actions/test_escalate_action.py
@@ -252,7 +252,9 @@ class TestEscalateAction:
     )
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPathConfig")
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPath")
-    @patch("uipath_langchain.agent.tools.escalation_tool.resolve_recipient_value")
+    @patch(
+        "uipath_langchain.agent.guardrails.actions.escalate_action.resolve_recipient_value"
+    )
     async def test_create_task_node_sends_correct_data(
         self,
         mock_resolve_recipient,
@@ -330,7 +332,9 @@ class TestEscalateAction:
     @pytest.mark.asyncio
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPathConfig")
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPath")
-    @patch("uipath_langchain.agent.tools.escalation_tool.resolve_recipient_value")
+    @patch(
+        "uipath_langchain.agent.guardrails.actions.escalate_action.resolve_recipient_value"
+    )
     async def test_create_task_node_post_agent_with_agent_result(
         self,
         mock_resolve_recipient,
@@ -404,7 +408,9 @@ class TestEscalateAction:
             GuardrailScope.TOOL,
         ],
     )
-    @patch("uipath_langchain.agent.tools.escalation_tool.resolve_recipient_value")
+    @patch(
+        "uipath_langchain.agent.guardrails.actions.escalate_action.resolve_recipient_value"
+    )
     async def test_create_task_node_post_execution_single_message_raises_error(
         self, mock_resolve_recipient, scope: GuardrailScope
     ):
@@ -437,7 +443,9 @@ class TestEscalateAction:
     @pytest.mark.asyncio
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPathConfig")
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPath")
-    @patch("uipath_langchain.agent.tools.escalation_tool.resolve_recipient_value")
+    @patch(
+        "uipath_langchain.agent.guardrails.actions.escalate_action.resolve_recipient_value"
+    )
     async def test_create_task_node_tool_pre_execution_extracts_tool_args(
         self,
         mock_resolve_recipient,
@@ -497,7 +505,9 @@ class TestEscalateAction:
     @pytest.mark.asyncio
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPathConfig")
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPath")
-    @patch("uipath_langchain.agent.tools.escalation_tool.resolve_recipient_value")
+    @patch(
+        "uipath_langchain.agent.guardrails.actions.escalate_action.resolve_recipient_value"
+    )
     async def test_create_task_node_tool_post_execution_extracts_tool_content(
         self,
         mock_resolve_recipient,
@@ -559,7 +569,9 @@ class TestEscalateAction:
     @pytest.mark.asyncio
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPathConfig")
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPath")
-    @patch("uipath_langchain.agent.tools.escalation_tool.resolve_recipient_value")
+    @patch(
+        "uipath_langchain.agent.guardrails.actions.escalate_action.resolve_recipient_value"
+    )
     async def test_create_task_node_post_execution_ai_message_with_tool_calls(
         self,
         mock_resolve_recipient,
@@ -1730,7 +1742,9 @@ class TestEscalateAction:
     )
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPathConfig")
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPath")
-    @patch("uipath_langchain.agent.tools.escalation_tool.resolve_recipient_value")
+    @patch(
+        "uipath_langchain.agent.guardrails.actions.escalate_action.resolve_recipient_value"
+    )
     async def test_create_task_resolves_recipient_correctly(
         self,
         mock_resolve_recipient,
@@ -1772,7 +1786,9 @@ class TestEscalateAction:
         assert call_kwargs["recipient"] == expected_value
 
     @pytest.mark.asyncio
-    @patch("uipath_langchain.agent.tools.escalation_tool.resolve_recipient_value")
+    @patch(
+        "uipath_langchain.agent.guardrails.actions.escalate_action.resolve_recipient_value"
+    )
     async def test_create_task_with_asset_recipient_resolution_failure(
         self, mock_resolve_recipient
     ) -> None:
@@ -1850,7 +1866,9 @@ class TestEscalateActionMetadata:
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPathConfig")
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPath")
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.interrupt")
-    @patch("uipath_langchain.agent.tools.escalation_tool.resolve_recipient_value")
+    @patch(
+        "uipath_langchain.agent.guardrails.actions.escalate_action.resolve_recipient_value"
+    )
     async def test_standard_recipient_assigned_to_uses_value(
         self, mock_resolve_recipient, mock_interrupt, mock_uipath_class, mock_config
     ):
@@ -1892,7 +1910,9 @@ class TestEscalateActionMetadata:
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPathConfig")
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPath")
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.interrupt")
-    @patch("uipath_langchain.agent.tools.escalation_tool.resolve_recipient_value")
+    @patch(
+        "uipath_langchain.agent.guardrails.actions.escalate_action.resolve_recipient_value"
+    )
     async def test_standard_recipient_assigned_to_uses_display_name(
         self, mock_resolve_recipient, mock_interrupt, mock_uipath_class, mock_config
     ):
@@ -1934,7 +1954,9 @@ class TestEscalateActionMetadata:
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPathConfig")
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPath")
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.interrupt")
-    @patch("uipath_langchain.agent.tools.escalation_tool.resolve_recipient_value")
+    @patch(
+        "uipath_langchain.agent.guardrails.actions.escalate_action.resolve_recipient_value"
+    )
     async def test_asset_recipient_assigned_to_uses_resolved_value(
         self, mock_resolve_recipient, mock_interrupt, mock_uipath_class, mock_config
     ):
@@ -1998,7 +2020,9 @@ class TestEscalateActionMetadata:
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPathConfig")
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPath")
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.interrupt")
-    @patch("uipath_langchain.agent.tools.escalation_tool.resolve_recipient_value")
+    @patch(
+        "uipath_langchain.agent.guardrails.actions.escalate_action.resolve_recipient_value"
+    )
     async def test_workload_and_roundrobin_assigned_to_uses_display_name(
         self,
         mock_resolve_recipient,
@@ -2046,7 +2070,9 @@ class TestEscalateActionMetadata:
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPathConfig")
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPath")
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.interrupt")
-    @patch("uipath_langchain.agent.tools.escalation_tool.resolve_recipient_value")
+    @patch(
+        "uipath_langchain.agent.guardrails.actions.escalate_action.resolve_recipient_value"
+    )
     async def test_custom_assignees_assigned_to_uses_display_name(
         self, mock_resolve_recipient, mock_interrupt, mock_uipath_class, mock_config
     ):
@@ -2092,7 +2118,9 @@ class TestEscalateActionMetadata:
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPathConfig")
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPath")
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.interrupt")
-    @patch("uipath_langchain.agent.tools.escalation_tool.resolve_recipient_value")
+    @patch(
+        "uipath_langchain.agent.guardrails.actions.escalate_action.resolve_recipient_value"
+    )
     async def test_custom_assignees_assigned_to_falls_back_to_value(
         self, mock_resolve_recipient, mock_interrupt, mock_uipath_class, mock_config
     ):
@@ -2136,7 +2164,9 @@ class TestEscalateActionMetadata:
     @pytest.mark.asyncio
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPathConfig")
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPath")
-    @patch("uipath_langchain.agent.tools.escalation_tool.resolve_recipient_value")
+    @patch(
+        "uipath_langchain.agent.guardrails.actions.escalate_action.resolve_recipient_value"
+    )
     async def test_create_task_node_sets_metadata_node_type(
         self, mock_resolve_recipient, mock_uipath_class, mock_config
     ):
