+++ tests/agent/tools/test_escalation_tool.py
@@ -4,39 +4,25 @@ import os
 from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
-from langchain_core.messages import AIMessage, HumanMessage, ToolCall, ToolMessage
+from langchain_core.messages import ToolCall
 from uipath.agent.models.agent import (
     AgentEscalationChannel,
     AgentEscalationChannelProperties,
-    AgentEscalationRecipient,
     AgentEscalationRecipientType,
     AgentEscalationResourceConfig,
     AgentQuickFormChannelProperties,
     AgentQuickFormEscalationChannel,
-    ArgumentEmailRecipient,
-    ArgumentGroupNameRecipient,
-    AssetRecipient,
-    CustomAssigneesRecipient,
-    RoundRobinRecipient,
     StandardRecipient,
-    ToolOutputRecipient,
-    WorkloadRecipient,
 )
 from uipath.platform.action_center.tasks import Task, TaskRecipient, TaskRecipientType
 
 from uipath_langchain.agent.tools.escalation_memory import (
     EscalationMemoryCachedResult,
-    _get_user_email,
 )
 from uipath_langchain.agent.tools.escalation_tool import (
     _build_escalation_memory_payload,
-    _build_tool_output_task_recipient,
-    _extract_tool_output_value,
     _parse_task_data,
     create_escalation_tool,
-    resolve_asset,
-    resolve_channel_recipients,
-    resolve_recipient_value,
 )
 
 
@@ -45,358 +31,6 @@ def _make_mock_task(**overrides):
     defaults = {"id": 1, "key": "task-key", "title": "Test Task"}
     defaults.update(overrides)
     return Task(**defaults)
-
-
-class TestResolveAsset:
-    """Test the resolve_asset function."""
-
-    @pytest.mark.asyncio
-    @patch("uipath_langchain.agent.tools.escalation_tool.UiPath")
-    async def test_resolve_asset_success(self, mock_uipath_class):
-        """Test successful asset retrieval."""
-        # Setup mock
-        mock_client = MagicMock()
-        mock_uipath_class.return_value = mock_client
-        mock_result = MagicMock()
-        mock_result.value = "test@example.com"
-        mock_client.assets.retrieve_async = AsyncMock(return_value=mock_result)
-
-        # Execute
-        result = await resolve_asset("email_asset", "/Test/Folder")
-
-        # Assert
-        assert result == "test@example.com"
-        mock_client.assets.retrieve_async.assert_called_once_with(
-            name="email_asset", folder_path="/Test/Folder"
-        )
-
-    @pytest.mark.asyncio
-    @patch("uipath_langchain.agent.tools.escalation_tool.UiPath")
-    async def test_resolve_asset_no_value(self, mock_uipath_class):
-        """Test asset with no value raises ValueError."""
-        # Setup mock
-        mock_client = MagicMock()
-        mock_uipath_class.return_value = mock_client
-        mock_result = MagicMock()
-        mock_result.value = None
-        mock_client.assets.retrieve_async = AsyncMock(return_value=mock_result)
-
-        # Execute and assert
-        with pytest.raises(ValueError) as exc_info:
-            await resolve_asset("empty_asset", "/Test/Folder")
-
-        assert "Asset 'empty_asset' has no value configured" in str(exc_info.value)
-
-    @pytest.mark.asyncio
-    @patch("uipath_langchain.agent.tools.escalation_tool.UiPath")
-    async def test_resolve_asset_not_found(self, mock_uipath_class):
-        """Test asset not found raises ValueError."""
-        # Setup mock
-        mock_client = MagicMock()
-        mock_uipath_class.return_value = mock_client
-        mock_client.assets.retrieve_async = AsyncMock(return_value=None)
-
-        # Execute and assert
-        with pytest.raises(ValueError) as exc_info:
-            await resolve_asset("missing_asset", "/Test/Folder")
-
-        assert "Asset 'missing_asset' has no value configured" in str(exc_info.value)
-
-    @pytest.mark.asyncio
-    @patch("uipath_langchain.agent.tools.escalation_tool.UiPath")
-    async def test_resolve_asset_retrieval_exception(self, mock_uipath_class):
-        """Test exception during asset retrieval raises ValueError with context."""
-        # Setup mock
-        mock_client = MagicMock()
-        mock_uipath_class.return_value = mock_client
-        mock_client.assets.retrieve_async = AsyncMock(
-            side_effect=Exception("Connection error")
-        )
-
-        # Execute and assert
-        with pytest.raises(ValueError) as exc_info:
-            await resolve_asset("problem_asset", "/Test/Folder")
-
-        assert (
-            "Failed to resolve asset 'problem_asset' in folder '/Test/Folder'"
-            in str(exc_info.value)
-        )
-        assert "Connection error" in str(exc_info.value)
-
-
-class TestResolveRecipientValue:
-    """Test the resolve_recipient_value function."""
-
-    @pytest.mark.asyncio
-    @patch.dict(os.environ, {"UIPATH_FOLDER_PATH": "/Test/Folder"})
-    @patch("uipath_langchain.agent.tools.escalation_tool.resolve_asset")
-    async def test_resolve_recipient_asset_user_email(self, mock_resolve_asset):
-        """Test ASSET_USER_EMAIL type calls resolve_asset."""
-        mock_resolve_asset.return_value = "resolved@example.com"
-
-        recipient = AssetRecipient(
-            type=AgentEscalationRecipientType.ASSET_USER_EMAIL,
-            asset_name="email_asset",
-            folder_path="/Test/Folder",
-        )
-
-        result = await resolve_recipient_value(recipient)
-
-        assert result == TaskRecipient(
-            value="resolved@example.com",
-            type=TaskRecipientType.EMAIL,
-            displayName="resolved@example.com",
-        )
-        mock_resolve_asset.assert_called_once_with("email_asset", "/Test/Folder")
-
-    @pytest.mark.asyncio
-    @patch.dict(os.environ, {"UIPATH_FOLDER_PATH": "/Test/Folder"})
-    @patch("uipath_langchain.agent.tools.escalation_tool.resolve_asset")
-    async def test_resolve_recipient_asset_group_name(self, mock_resolve_asset):
-        """Test ASSET_GROUP_NAME type calls resolve_asset."""
-        mock_resolve_asset.return_value = "ResolvedGroup"
-
-        recipient = AssetRecipient(
-            type=AgentEscalationRecipientType.ASSET_GROUP_NAME,
-            asset_name="group_asset",
-            folder_path="/Test/Folder",
-        )
-
-        result = await resolve_recipient_value(recipient)
-
-        assert result == TaskRecipient(
-            value="ResolvedGroup",
-            type=TaskRecipientType.GROUP_NAME,
-            displayName="ResolvedGroup",
-        )
-        mock_resolve_asset.assert_called_once_with("group_asset", "/Test/Folder")
-
-    @pytest.mark.asyncio
-    async def test_resolve_recipient_user_email(self):
-        """Test USER_EMAIL type returns value directly."""
-        recipient = StandardRecipient(
-            type=AgentEscalationRecipientType.USER_EMAIL,
-            value="direct@example.com",
-        )
-
-        result = await resolve_recipient_value(recipient)
-
-        assert result == TaskRecipient(
-            value="direct@example.com",
-            type=TaskRecipientType.EMAIL,
-            displayName="direct@example.com",
-        )
-
-    @pytest.mark.asyncio
-    @patch("uipath_langchain.agent.tools.escalation_tool.resolve_asset")
-    async def test_resolve_recipient_propagates_error_when_asset_resolution_fails(
-        self, mock_resolve_asset
-    ):
-        """Test AssetRecipient when asset resolution fails."""
-        mock_resolve_asset.side_effect = ValueError("Asset not found")
-
-        recipient = AssetRecipient(
-            type=AgentEscalationRecipientType.ASSET_USER_EMAIL,
-            asset_name="nonexistent",
-            folder_path="Shared",
-        )
-
-        with pytest.raises(ValueError) as exc_info:
-            await resolve_recipient_value(recipient)
-
-        assert "Asset not found" in str(exc_info.value)
-
-    @pytest.mark.asyncio
-    async def test_resolve_recipient_no_value(self):
-        """Test recipient without value attribute returns None."""
-        # Create a minimal recipient object without value
-        recipient = MagicMock()
-        recipient.type = AgentEscalationRecipientType.USER_EMAIL
-        del recipient.value  # Simulate no value attribute
-
-        result = await resolve_recipient_value(recipient)
-
-        assert result is None
-
-    @pytest.mark.asyncio
-    async def test_resolve_recipient_workload(self):
-        """WorkloadRecipient resolves to TaskRecipient with WORKLOAD type using displayName."""
-        recipient = WorkloadRecipient(
-            type=AgentEscalationRecipientType.WORKLOAD,
-            value="group-id-1",
-            display_name="Support Team",
-        )
-
-        result = await resolve_recipient_value(recipient)
-
-        assert result == TaskRecipient(
-            value="Support Team",
-            type=TaskRecipientType.WORKLOAD,
-            displayName="Support Team",
-        )
-
-    @pytest.mark.asyncio
-    async def test_resolve_recipient_round_robin(self):
-        """RoundRobinRecipient resolves to TaskRecipient with ROUND_ROBIN type."""
-        recipient = RoundRobinRecipient(
-            type=AgentEscalationRecipientType.ROUND_ROBIN,
-            value="group-id-1",
-            display_name="Support Team",
-        )
-
-        result = await resolve_recipient_value(recipient)
-
-        assert result == TaskRecipient(
-            value="Support Team",
-            type=TaskRecipientType.ROUND_ROBIN,
-            displayName="Support Team",
-        )
-
-    @pytest.mark.asyncio
-    async def test_resolve_recipient_custom_assignees_single(self):
-        """A single CustomAssigneesRecipient becomes a one-element Workload assignment."""
-        recipient = CustomAssigneesRecipient(
-            type=AgentEscalationRecipientType.CUSTOM_ASSIGNEES,
-            value="alice@example.com",
-            displayName="Alice",
-        )
-
-        result = await resolve_recipient_value(recipient)
-
-        assert result == TaskRecipient(
-            value="alice@example.com",
-            values=["alice@example.com"],
-            type=TaskRecipientType.WORKLOAD,
-            displayName="Alice",
-        )
-
-    @pytest.mark.asyncio
-    async def test_resolve_recipient_custom_assignees_empty_value_returns_none(self):
-        """Empty-value CustomAssignees sentinel resolves to None."""
-        recipient = CustomAssigneesRecipient(
-            type=AgentEscalationRecipientType.CUSTOM_ASSIGNEES,
-            value="",
-        )
-
-        result = await resolve_recipient_value(recipient)
-
-        assert result is None
-
-
-class TestResolveChannelRecipients:
-    """Tests for the channel-level recipient aggregator."""
-
-    @pytest.mark.asyncio
-    async def test_empty_recipients_returns_none(self):
-        result = await resolve_channel_recipients([])
-        assert result is None
-
-    @pytest.mark.asyncio
-    async def test_single_workload_recipient_delegates_to_resolve_recipient_value(self):
-        recipient = WorkloadRecipient(
-            type=AgentEscalationRecipientType.WORKLOAD,
-            value="group-1",
-            display_name="Support Team",
-        )
-
-        result = await resolve_channel_recipients([recipient])
-
-        assert result == TaskRecipient(
-            value="Support Team",
-            type=TaskRecipientType.WORKLOAD,
-            displayName="Support Team",
-        )
-
-    @pytest.mark.asyncio
-    async def test_multiple_custom_assignees_collected_into_single_workload(self):
-        """Multiple CustomAssignees recipients aggregate into one Workload + values list."""
-        recipients: list[AgentEscalationRecipient] = [
-            CustomAssigneesRecipient(
-                type=AgentEscalationRecipientType.CUSTOM_ASSIGNEES,
-                value="alice@example.com",
-                displayName="Alice",
-            ),
-            CustomAssigneesRecipient(
-                type=AgentEscalationRecipientType.CUSTOM_ASSIGNEES,
-                value="bob@example.com",
-                displayName="Bob",
-            ),
-        ]
-
-        result = await resolve_channel_recipients(recipients)
-
-        assert result == TaskRecipient(
-            value="alice@example.com",
-            values=["alice@example.com", "bob@example.com"],
-            type=TaskRecipientType.WORKLOAD,
-        )
-
-    @pytest.mark.asyncio
-    async def test_single_custom_assignee_collected_into_single_workload(self):
-        recipients: list[AgentEscalationRecipient] = [
-            CustomAssigneesRecipient(
-                type=AgentEscalationRecipientType.CUSTOM_ASSIGNEES,
-                value="alice@example.com",
-            ),
-        ]
-
-        result = await resolve_channel_recipients(recipients)
-
-        assert result == TaskRecipient(
-            value="alice@example.com",
-            values=["alice@example.com"],
-            type=TaskRecipientType.WORKLOAD,
-        )
-
-    @pytest.mark.asyncio
-    async def test_custom_assignees_filters_empty_sentinel_values(self):
-        """Empty-value sentinel CustomAssignees are skipped, not included in the array."""
-        recipients: list[AgentEscalationRecipient] = [
-            CustomAssigneesRecipient(
-                type=AgentEscalationRecipientType.CUSTOM_ASSIGNEES,
-                value="",
-            ),
-            CustomAssigneesRecipient(
-                type=AgentEscalationRecipientType.CUSTOM_ASSIGNEES,
-                value="alice@example.com",
-            ),
-        ]
-
-        result = await resolve_channel_recipients(recipients)
-
-        assert result == TaskRecipient(
-            value="alice@example.com",
-            values=["alice@example.com"],
-            type=TaskRecipientType.WORKLOAD,
-        )
-
-    @pytest.mark.asyncio
-    async def test_all_empty_custom_assignees_returns_none(self):
-        recipients: list[AgentEscalationRecipient] = [
-            CustomAssigneesRecipient(
-                type=AgentEscalationRecipientType.CUSTOM_ASSIGNEES,
-                value="",
-            ),
-        ]
-
-        result = await resolve_channel_recipients(recipients)
-        assert result is None
-
-    @pytest.mark.asyncio
-    async def test_standard_recipient_uses_first_recipient_only(self):
-        """Non-CustomAssignees channels only use recipients[0]."""
-        recipient = StandardRecipient(
-            type=AgentEscalationRecipientType.USER_EMAIL,
-            value="alice@example.com",
-        )
-
-        result = await resolve_channel_recipients([recipient])
-
-        assert result == TaskRecipient(
-            value="alice@example.com",
-            type=TaskRecipientType.EMAIL,
-            displayName="alice@example.com",
-        )
 
 
 class TestEscalationToolMetadata:
@@ -1048,34 +682,6 @@ class TestEscalationToolOutputSchema:
             mock_check_memory_cache.await_args.kwargs["memory_space_name"]
             == "MemorySpace"
         )
-
-
-class TestGetUserEmail:
-    """Test the _get_user_email helper function."""
-
-    def test_none_returns_none(self):
-        """Test that None input returns None."""
-        assert _get_user_email(None) is None
-
-    def test_dict_with_email_address(self):
-        """Test extraction from dict with emailAddress field."""
-        user = {"emailAddress": "test@example.com", "name": "Test"}
-        assert _get_user_email(user) == "test@example.com"
-
-    def test_dict_without_email_address(self):
-        """Test dict without emailAddress returns None."""
-        user = {"name": "Test", "id": 123}
-        assert _get_user_email(user) is None
-
-    def test_object_with_email_address(self):
-        """Test extraction from object with emailAddress attribute."""
-        user = MagicMock(emailAddress="test@example.com")
-        assert _get_user_email(user) == "test@example.com"
-
-    def test_object_without_email_address(self):
-        """Test object without emailAddress attribute returns None."""
-        user = MagicMock(spec=["name", "id"])
-        assert _get_user_email(user) is None
 
 
 class TestEscalationToolTaskInfo:
@@ -1760,583 +1366,3 @@ class TestQuickFormEscalation:
 
         mock_client.tasks.create_async.assert_called_once()
         mock_client.tasks.create_quickform_async.assert_not_called()
-
-
-class TestExtractToolOutputValue:
-    """Tests for the helper that walks message history to extract tool output values."""
-
-    def test_extracts_top_level_field_from_json_content(self):
-        messages = [
-            HumanMessage(content="hello"),
-            ToolMessage(
-                name="API workflow A",
-                tool_call_id="1",
-                content='{"includeEmails": ["alice@x.com", "bob@x.com"], "expiresAt": "2026-01-01"}',
-            ),
-        ]
-        result = _extract_tool_output_value(messages, "API workflow A", "includeEmails")
-        assert result == ["alice@x.com", "bob@x.com"]
-
-    def test_returns_whole_parsed_content_when_path_is_empty(self):
-        messages = [
-            ToolMessage(
-                name="getApprovers",
-                tool_call_id="1",
-                content='["alice@x.com", "bob@x.com"]',
-            ),
-        ]
-        result = _extract_tool_output_value(messages, "getApprovers", "")
-        assert result == ["alice@x.com", "bob@x.com"]
-
-    def test_picks_latest_invocation_when_called_multiple_times(self):
-        messages = [
-            ToolMessage(
-                name="A", tool_call_id="1", content='{"emails": ["old@x.com"]}'
-            ),
-            ToolMessage(
-                name="A", tool_call_id="2", content='{"emails": ["new@x.com"]}'
-            ),
-        ]
-        result = _extract_tool_output_value(messages, "A", "emails")
-        assert result == ["new@x.com"]
-
-    def test_raises_when_tool_was_never_called(self):
-        messages = [
-            ToolMessage(name="getUsers", tool_call_id="1", content="[]"),
-        ]
-        with pytest.raises(ValueError, match="has not been called yet"):
-            _extract_tool_output_value(messages, "API workflow A", "emails")
-
-    def test_raises_when_field_missing_from_output(self):
-        messages = [
-            ToolMessage(name="A", tool_call_id="1", content='{"otherField": []}'),
-        ]
-        with pytest.raises(ValueError, match="does not contain field 'emails'"):
-            _extract_tool_output_value(messages, "A", "emails")
-
-    def test_raises_when_output_is_not_a_json_object_but_path_requested(self):
-        messages = [
-            ToolMessage(name="A", tool_call_id="1", content='["a", "b"]'),
-        ]
-        with pytest.raises(ValueError, match="is not a JSON object"):
-            _extract_tool_output_value(messages, "A", "emails")
-
-    def test_handles_non_json_string_content_gracefully(self):
-        # If the content isn't JSON, treat it as a raw string (the whole "output").
-        messages = [
-            ToolMessage(name="A", tool_call_id="1", content="some raw text"),
-        ]
-        # With no path, the raw string is returned.
-        result = _extract_tool_output_value(messages, "A", "")
-        assert result == "some raw text"
-
-    def test_ignores_non_matching_tool_messages(self):
-        messages = [
-            ToolMessage(
-                name="otherTool", tool_call_id="1", content='{"emails": ["wrong"]}'
-            ),
-            ToolMessage(
-                name="A", tool_call_id="2", content='{"emails": ["right@x.com"]}'
-            ),
-            AIMessage(content="thinking"),
-        ]
-        result = _extract_tool_output_value(messages, "A", "emails")
-        assert result == ["right@x.com"]
-
-    def test_returns_empty_string_when_content_is_empty_and_no_path(self):
-        # Empty content with no path falls through the JSON parse to the raw "" string.
-        messages = [ToolMessage(name="A", tool_call_id="1", content="")]
-        result = _extract_tool_output_value(messages, "A", "")
-        assert result == ""
-
-    def test_raises_when_malformed_json_and_path_requested(self):
-        # Malformed JSON content falls back to the raw string; attempting to extract a
-        # path then fails the same way as any other non-object output.
-        messages = [
-            ToolMessage(name="A", tool_call_id="1", content='{"emails": ["a@x.com"'),
-        ]
-        with pytest.raises(ValueError, match="is not a JSON object"):
-            _extract_tool_output_value(messages, "A", "emails")
-
-    def test_returns_null_when_field_value_is_explicitly_null(self):
-        # A null field value at the requested path is returned as Python None; the
-        # downstream recipient builder is the one that rejects None/empty values.
-        messages = [
-            ToolMessage(name="A", tool_call_id="1", content='{"emails": null}'),
-        ]
-        result = _extract_tool_output_value(messages, "A", "emails")
-        assert result is None
-
-    def test_uses_non_string_content_directly_without_parsing(self):
-        # When ToolMessage.content is already structured (e.g. dict), use it
-        # as-is rather than attempting json.loads on a non-string.
-        messages = [
-            ToolMessage(
-                name="A",
-                tool_call_id="1",
-                content=[{"type": "text", "text": '{"emails": ["a@x.com"]}'}],
-            ),
-        ]
-        # Non-string content is returned verbatim when no path is requested.
-        result = _extract_tool_output_value(messages, "A", "")
-        assert result == [{"type": "text", "text": '{"emails": ["a@x.com"]}'}]
-
-
-class TestBuildToolOutputTaskRecipient:
-    """Tests for _build_tool_output_task_recipient single-value & edge branches."""
-
-    def test_returns_group_id_recipient_for_group_id_criteria(self):
-        # Single string mapped through the GROUP_ID branch.
-        result = _build_tool_output_task_recipient(
-            AgentEscalationRecipientType.GROUP_ID, "group-123"
-        )
-        assert result is not None
-        assert result.value == "group-123"
-        assert result.type == TaskRecipientType.GROUP_ID
-
-    def test_returns_user_id_recipient_for_user_id_criteria(self):
-        result = _build_tool_output_task_recipient(
-            AgentEscalationRecipientType.USER_ID, "user-1"
-        )
-        assert result is not None
-        assert result.value == "user-1"
-        assert result.type == TaskRecipientType.USER_ID
-
-    def test_workload_string_single_value(self):
-        result = _build_tool_output_task_recipient(
-            AgentEscalationRecipientType.WORKLOAD, "workload-1"
-        )
-        assert result is not None
-        assert result.values == ["workload-1"]
-        assert result.type == TaskRecipientType.WORKLOAD
-
-    def test_round_robin_string_single_value(self):
-        result = _build_tool_output_task_recipient(
-            AgentEscalationRecipientType.ROUND_ROBIN, "rr-1"
-        )
-        assert result is not None
-        assert result.type == TaskRecipientType.ROUND_ROBIN
-
-    def test_custom_assignees_splits_comma_separated_string(self):
-        result = _build_tool_output_task_recipient(
-            AgentEscalationRecipientType.CUSTOM_ASSIGNEES,
-            "alice@example.com, bob@example.com , ,",
-        )
-        assert result is not None
-        assert result.values == ["alice@example.com", "bob@example.com"]
-        assert result.type == TaskRecipientType.WORKLOAD
-
-    def test_custom_assignees_returns_none_for_only_separators(self):
-        # A string with only commas/whitespace yields no valid emails -> None.
-        result = _build_tool_output_task_recipient(
-            AgentEscalationRecipientType.CUSTOM_ASSIGNEES, ", , ,"
-        )
-        assert result is None
-
-    def test_returns_none_for_unknown_criteria_type(self):
-        # Recipient types that aren't single-valued criteria (e.g. an asset
-        # type accidentally routed through here) fall through to a None.
-        result = _build_tool_output_task_recipient(
-            AgentEscalationRecipientType.ASSET_USER_EMAIL, "value"
-        )
-        assert result is None
-
-    def test_raises_when_resolved_string_value_is_empty(self):
-        with pytest.raises(ValueError, match="empty value"):
-            _build_tool_output_task_recipient(AgentEscalationRecipientType.USER_ID, "")
-
-    def test_raises_when_resolved_value_is_none_string_coerce(self):
-        with pytest.raises(ValueError, match="empty value"):
-            _build_tool_output_task_recipient(
-                AgentEscalationRecipientType.WORKLOAD, None
-            )
-
-
-class TestArgumentRecipientResolutionMissing:
-    """Argument-name recipients raise loudly when the named input is missing."""
-
-    @pytest.mark.asyncio
-    async def test_argument_email_raises_when_input_missing(self):
-        recipient = ArgumentEmailRecipient(
-            type=AgentEscalationRecipientType.ARGUMENT_EMAIL,
-            argument_path="user.email",
-        )
-        with pytest.raises(ValueError, match="no value in agent input"):
-            await resolve_recipient_value(recipient, input_args={})
-
-    @pytest.mark.asyncio
-    async def test_argument_group_name_raises_when_input_missing(self):
-        recipient = ArgumentGroupNameRecipient(
-            type=AgentEscalationRecipientType.ARGUMENT_GROUP_NAME,
-            argument_path="group.name",
-        )
-        with pytest.raises(ValueError, match="no value in agent input"):
-            await resolve_recipient_value(recipient, input_args={})
-
-    @pytest.mark.asyncio
-    async def test_argument_email_resolves_value_from_input(self):
-        recipient = ArgumentEmailRecipient(
-            type=AgentEscalationRecipientType.ARGUMENT_EMAIL,
-            argument_path="approver.email",
-        )
-        result = await resolve_recipient_value(
-            recipient,
-            input_args={"approver": {"email": "boss@example.com"}},
-        )
-        assert result is not None
-        assert result.value == "boss@example.com"
-        assert result.type == TaskRecipientType.EMAIL
-
-    @pytest.mark.asyncio
-    async def test_argument_group_name_resolves_value_from_input(self):
-        recipient = ArgumentGroupNameRecipient(
-            type=AgentEscalationRecipientType.ARGUMENT_GROUP_NAME,
-            argument_path="dept.team",
-        )
-        result = await resolve_recipient_value(
-            recipient,
-            input_args={"dept": {"team": "Finance"}},
-        )
-        assert result is not None
-        assert result.value == "Finance"
-        assert result.type == TaskRecipientType.GROUP_NAME
-
-
-class TestToolOutputRecipientResolution:
-    """Tests for resolving ToolOutputRecipient via resolve_recipient_value."""
-
-    @pytest.mark.asyncio
-    async def test_custom_assignees_with_array_of_emails(self):
-        recipient = ToolOutputRecipient(
-            type=AgentEscalationRecipientType.CUSTOM_ASSIGNEES,
-            source="toolOutput",
-            tool_name="API workflow A",
-            output_path="includeEmails",
-        )
-        messages = [
-            ToolMessage(
-                name="API workflow A",
-                tool_call_id="1",
-                content='{"includeEmails": ["alice@x.com", "bob@x.com"]}',
-            ),
-        ]
-
-        result = await resolve_recipient_value(recipient, tool_messages=messages)
-
-        assert result == TaskRecipient(
-            value="alice@x.com",
-            values=["alice@x.com", "bob@x.com"],
-            type=TaskRecipientType.WORKLOAD,
-        )
-
-    @pytest.mark.asyncio
-    async def test_custom_assignees_with_comma_separated_string(self):
-        recipient = ToolOutputRecipient(
-            type=AgentEscalationRecipientType.CUSTOM_ASSIGNEES,
-            source="toolOutput",
-            tool_name="A",
-            output_path="emails",
-        )
-        messages = [
-            ToolMessage(
-                name="A",
-                tool_call_id="1",
-                content='{"emails": "alice@x.com, bob@x.com,  carol@x.com"}',
-            ),
-        ]
-
-        result = await resolve_recipient_value(recipient, tool_messages=messages)
-
-        assert result is not None
-        assert result.type == TaskRecipientType.WORKLOAD
-        assert result.values == ["alice@x.com", "bob@x.com", "carol@x.com"]
-
-    @pytest.mark.asyncio
-    async def test_workload_with_single_group_name_string(self):
-        recipient = ToolOutputRecipient(
-            type=AgentEscalationRecipientType.WORKLOAD,
-            source="toolOutput",
-            tool_name="findGroup",
-            output_path="groupName",
-        )
-        messages = [
-            ToolMessage(
-                name="findGroup",
-                tool_call_id="1",
-                content='{"groupName": "Support Team"}',
-            ),
-        ]
-
-        result = await resolve_recipient_value(recipient, tool_messages=messages)
-
-        assert result == TaskRecipient(
-            value="Support Team",
-            values=["Support Team"],
-            type=TaskRecipientType.WORKLOAD,
-        )
-
-    @pytest.mark.asyncio
-    async def test_round_robin_with_single_group_name(self):
-        recipient = ToolOutputRecipient(
-            type=AgentEscalationRecipientType.ROUND_ROBIN,
-            source="toolOutput",
-            tool_name="findGroup",
-            output_path="groupName",
-        )
-        messages = [
-            ToolMessage(
-                name="findGroup",
-                tool_call_id="1",
-                content='{"groupName": "Support Team"}',
-            ),
-        ]
-
-        result = await resolve_recipient_value(recipient, tool_messages=messages)
-
-        assert result == TaskRecipient(
-            value="Support Team",
-            values=["Support Team"],
-            type=TaskRecipientType.ROUND_ROBIN,
-        )
-
-    @pytest.mark.asyncio
-    async def test_user_with_string_value(self):
-        recipient = ToolOutputRecipient(
-            type=AgentEscalationRecipientType.USER_ID,
-            source="toolOutput",
-            tool_name="findUser",
-            output_path="userId",
-        )
-        messages = [
-            ToolMessage(
-                name="findUser",
-                tool_call_id="1",
-                content='{"userId": "user-123"}',
-            ),
-        ]
-
-        result = await resolve_recipient_value(recipient, tool_messages=messages)
-
-        assert result == TaskRecipient(value="user-123", type=TaskRecipientType.USER_ID)
-
-    @pytest.mark.asyncio
-    async def test_raises_when_tool_not_called(self):
-        recipient = ToolOutputRecipient(
-            type=AgentEscalationRecipientType.CUSTOM_ASSIGNEES,
-            source="toolOutput",
-            tool_name="API workflow A",
-            output_path="includeEmails",
-        )
-        with pytest.raises(ValueError, match="has not been called yet"):
-            await resolve_recipient_value(recipient, tool_messages=[])
-
-    @pytest.mark.asyncio
-    async def test_raises_when_resolved_value_is_empty_list(self):
-        recipient = ToolOutputRecipient(
-            type=AgentEscalationRecipientType.CUSTOM_ASSIGNEES,
-            source="toolOutput",
-            tool_name="A",
-            output_path="emails",
-        )
-        messages = [
-            ToolMessage(name="A", tool_call_id="1", content='{"emails": []}'),
-        ]
-        with pytest.raises(ValueError, match="empty list"):
-            await resolve_recipient_value(recipient, tool_messages=messages)
-
-    @pytest.mark.asyncio
-    @pytest.mark.parametrize(
-        "criteria",
-        [
-            AgentEscalationRecipientType.USER_ID,
-            AgentEscalationRecipientType.GROUP_ID,
-            AgentEscalationRecipientType.ROUND_ROBIN,
-        ],
-    )
-    async def test_raises_when_list_resolved_for_single_valued_criteria(self, criteria):
-        # Single-valued criteria (User/Group/RoundRobin) cannot accept a list of
-        # recipients; silently demoting to WORKLOAD would change the assignment
-        # semantics. The resolver should raise instead.
-        recipient = ToolOutputRecipient(
-            type=criteria,
-            source="toolOutput",
-            tool_name="A",
-            output_path="ids",
-        )
-        messages = [
-            ToolMessage(
-                name="A",
-                tool_call_id="1",
-                content='{"ids": ["one", "two"]}',
-            ),
-        ]
-        with pytest.raises(ValueError, match="expects a single value"):
-            await resolve_recipient_value(recipient, tool_messages=messages)
-
-    @pytest.mark.asyncio
-    async def test_tool_name_matches_sanitized_form(self):
-        # ToolMessage.name carries the sanitized tool name (matches the form
-        # used at tool-call time). Recipient configs may carry either the
-        # display name or the sanitized form — both must resolve.
-        from uipath_langchain.agent.tools.utils import sanitize_tool_name
-
-        display_name = "Get Users Info"
-        sanitized = sanitize_tool_name(display_name)
-        assert sanitized != display_name  # sanity check: forms differ
-
-        recipient = ToolOutputRecipient(
-            type=AgentEscalationRecipientType.USER_ID,
-            source="toolOutput",
-            tool_name=display_name,
-            output_path="userId",
-        )
-        messages = [
-            ToolMessage(name=sanitized, tool_call_id="1", content='{"userId": "u-1"}'),
-        ]
-
-        result = await resolve_recipient_value(recipient, tool_messages=messages)
-        assert result is not None
-        assert result.value == "u-1"
-        assert result.type == TaskRecipientType.USER_ID
-
-
-class TestResolveChannelRecipientsWithToolOutput:
-    """Tests for resolve_channel_recipients with tool-output bindings."""
-
-    @pytest.mark.asyncio
-    async def test_tool_output_recipient_delegates_to_resolver(self):
-        recipient = ToolOutputRecipient(
-            type=AgentEscalationRecipientType.CUSTOM_ASSIGNEES,
-            source="toolOutput",
-            tool_name="A",
-            output_path="emails",
-        )
-        messages = [
-            ToolMessage(
-                name="A",
-                tool_call_id="1",
-                content='{"emails": ["a@b.com", "c@d.com"]}',
-            ),
-        ]
-
-        result = await resolve_channel_recipients([recipient], tool_messages=messages)
-
-        assert result is not None
-        assert result.values == ["a@b.com", "c@d.com"]
-
-    @pytest.mark.asyncio
-    async def test_tool_output_takes_precedence_over_custom_assignees_aggregation(self):
-        # Even if subsequent recipients are CustomAssigneesRecipient, a leading
-        # ToolOutputRecipient owns the entire channel resolution.
-        recipients: list[AgentEscalationRecipient] = [
-            ToolOutputRecipient(
-                type=AgentEscalationRecipientType.CUSTOM_ASSIGNEES,
-                source="toolOutput",
-                tool_name="A",
-                output_path="emails",
-            ),
-            CustomAssigneesRecipient(
-                type=AgentEscalationRecipientType.CUSTOM_ASSIGNEES,
-                value="manual@x.com",
-            ),
-        ]
-        messages = [
-            ToolMessage(
-                name="A",
-                tool_call_id="1",
-                content='{"emails": ["resolved@x.com"]}',
-            ),
-        ]
-
-        result = await resolve_channel_recipients(recipients, tool_messages=messages)
-
-        assert result is not None
-        # Tool-output wins; manual recipient is ignored.
-        assert result.values == ["resolved@x.com"]
-
-
-class TestEscalationToolDescriptionAugmentation:
-    """Tests for the LLM-facing description hint added when tool-output bindings exist."""
-
-    def _make_channel(
-        self, recipients: list[AgentEscalationRecipient]
-    ) -> AgentEscalationChannel:
-        return AgentEscalationChannel(
-            id="ch-1",
-            type="actionCenter",
-            name="ch",
-            description="",
-            input_schema={"type": "object", "properties": {}},
-            output_schema={"type": "object", "properties": {}},
-            recipients=recipients,
-            properties=AgentEscalationChannelProperties(
-                app_name="my-app",
-                resource_key="rk",
-                folder_name=None,
-                app_version=1,
-                is_actionable_message_enabled=False,
-                actionable_message_meta_data=None,
-            ),
-        )
-
-    def _make_resource(
-        self, channel: AgentEscalationChannel
-    ) -> AgentEscalationResourceConfig:
-        return AgentEscalationResourceConfig(
-            resourceType="escalation",
-            id="esc-1",
-            name="approve_expense",
-            description="Escalate an expense for approval.",
-            channels=[channel],
-        )
-
-    def test_description_unchanged_when_no_tool_output_bindings(self):
-        channel = self._make_channel(
-            recipients=[
-                StandardRecipient(
-                    type=AgentEscalationRecipientType.USER_ID,
-                    value="u1",
-                    displayName="User 1",
-                )
-            ]
-        )
-        tool = create_escalation_tool(self._make_resource(channel))
-        assert tool.description == "Escalate an expense for approval."
-        assert "Recipient routing notes" not in tool.description
-
-    def test_description_includes_dependency_hint_for_tool_output_binding(self):
-        channel = self._make_channel(
-            recipients=[
-                ToolOutputRecipient(
-                    type=AgentEscalationRecipientType.CUSTOM_ASSIGNEES,
-                    source="toolOutput",
-                    tool_name="API workflow A",
-                    output_path="includeEmails",
-                )
-            ]
-        )
-        tool = create_escalation_tool(self._make_resource(channel))
-        assert "Recipient routing notes" in tool.description
-        assert "API workflow A" in tool.description
-        assert "includeEmails" in tool.description
-
-    def test_description_deduplicates_repeated_tool_dependencies(self):
-        channel = self._make_channel(
-            recipients=[
-                ToolOutputRecipient(
-                    type=AgentEscalationRecipientType.CUSTOM_ASSIGNEES,
-                    source="toolOutput",
-                    tool_name="A",
-                    output_path="emails",
-                ),
-                ToolOutputRecipient(
-                    type=AgentEscalationRecipientType.CUSTOM_ASSIGNEES,
-                    source="toolOutput",
-                    tool_name="A",
-                    output_path="emails",
-                ),
-            ]
-        )
-        tool = create_escalation_tool(self._make_resource(channel))
-        # The hint mentions tool A exactly once.
-        assert tool.description.count("`A`") == 1
